requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
elevenlabs==0.2.27
executing==2.0.1
frozenlist==1.4.1
gTTS==2.5.1
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
ipython==8.22.1
jedi==0.19.1
matplotlib-inline==0.1.6
multidict==6.0.5
numpy==1.26.4
openai==0.27.8
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
SpeechRecognition==3.10.1
stack-data==0.6.3
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
